Improve concurrency tests.
