+++ tests/compile-fail/concurrency/thread-spawn.rs
@@ -1,5 +1,6 @@
-// ignore-linux
-// ignore-macos
+// ignore-linux: Only Windows is not supported.
+// ignore-macos: Only Windows is not supported.
+
 use std::thread;
 
 // error-pattern: Miri does not support threading
