+++ tests/compile-fail/sync/libc_pthread_mutex_deadlock.rs
