+++ tests/run-pass/concurrency/simple.rs
@@ -1,4 +1,4 @@
-// ignore-windows
+// ignore-windows: Concurrency on Windows is not supported yet.
 
 use std::thread;
