+++ tests/run-pass/concurrency/thread_locals.stderr
@@ -1,2 +1,2 @@
-warning: The thread support is experimental. For example, Miri does not detect data races yet.
+warning: thread support is experimental. For example, Miri does not detect data races yet.
